Refactor: Add TUSChunk class for representing file chunks
This commit adds a new class called TUSChunk, which represents a chunk of a TUS file. A chunk is a part of a file that is uploaded to the server in a single request. The TUSChunk class has two member functions: getData() to retrieve the data of the chunk, and getChunkSize() to get the offset of the chunk in the file.

The addition of the TUSChunk class will enable more efficient handling of file uploads in the TUSClient library.