Refactor: Remove unused isLastRequestCompleted() function from IHttpClient
The isLastRequestCompleted() function in the IHttpClient interface is no longer used and has been removed. This function was previously used to check if the last request was completed, but it is no longer necessary for the current implementation.

This commit removes the isLastRequestCompleted() function from the IHttpClient interface to clean up the code and remove unused functionality.

Refactor: Remove unused isLastRequestCompleted() function from HttpClient

The isLastRequestCompleted() function in the HttpClient class is no longer used and has been removed. This function was previously used to check if the last request was completed, but it is no longer necessary for the current implementation.

This commit removes the isLastRequestCompleted() function from the HttpClient class to clean up the code and remove unused functionality.