Refactor: Update TusClient.cpp to fix resume() method
The resume() method in TusClient.cpp has been updated to fix a bug where the m_status was not properly set to READY before calling getUploadInfo(). This caused issues with the upload process. The code has been refactored to ensure that m_status is set correctly before calling getUploadInfo(). This commit resolves the issue and improves the reliability of the resume() method.